# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.Challenge:
    properties:
      category_id:
        type: integer
      created_at:
        type: string
      creator_id:
        type: string
      deadline_time:
        type: string
      description:
        type: string
      ends_at:
        type: string
      id:
        type: string
      invite_token:
        type: string
      is_public:
        type: boolean
      max_skips:
        type: integer
      starts_at:
        type: string
      status:
        type: string
      title:
        type: string
      working_days:
        items:
          type: integer
        type: array
    type: object
...
  domain.Progress:
    properties:
      adherence_pct:
        type: number
      checked_in_today:
        type: boolean
      current_streak:
        type: integer
      done_days:
        type: integer
      is_working_day:
        type: boolean
      max_streak:
        type: integer
      total_days:
        type: integer
    type: object
  domain.SimpleCheckIn:
    properties:
      challenge_id:
        type: string
      created_at:
        type: string
      date:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
...
  service.CreateChallengeParams:
    properties:
      category_id:
        type: integer
      deadline_time:
        type: string
      description:
        type: string
      ends_at:
        type: string
      is_public:
        type: boolean
      max_skips:
        type: integer
      starts_at:
        type: string
      title:
        type: string
      working_days:
        items:
          type: integer
        type: array
    type: object
  service.UpdateChallengeParams:
    properties:
      category_id:
        type: integer
      deadline_time:
        type: string
      description:
        type: string
      ends_at:
        type: string
      is_public:
        type: boolean
      max_skips:
        type: integer
      starts_at:
        type: string
      title:
        type: string
      working_days:
        items:
          type: integer
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  description: API для веб-сервиса коллективного трекинга привычек
  title: Habix API
  version: "1.0"
paths:
...
  /categories:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Category'
            type: array
      security:
      - BearerAuth: []
      summary: List all categories
      tags:
      - categories
  /challenges:
    get:
      parameters:
      - description: Filter public
        in: query
        name: public
        type: string
      - description: Category ID
        in: query
        name: category
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Challenge'
            type: array
      security:
      - BearerAuth: []
      summary: List public challenges
      tags:
      - challenges
    post:
      parameters:
      - description: Challenge data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.CreateChallengeParams'
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Challenge'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new challenge
      tags:
      - challenges
  /challenges/{id}:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Challenge'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get challenge by ID
      tags:
      - challenges
    patch:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/service.UpdateChallengeParams'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Challenge'
      security:
      - BearerAuth: []
      summary: Update a challenge
      tags:
      - challenges
  /challenges/{id}/checkin:
    delete:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Undo today's check-in
      tags:
      - checkin
    post:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.SimpleCheckIn'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Conflict
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Check in today
      tags:
      - checkin
  /challenges/{id}/checkins:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.SimpleCheckIn'
            type: array
      security:
      - BearerAuth: []
      summary: List all user check-ins in a challenge
      tags:
      - checkin
  /challenges/{id}/finish:
    post:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Finish a challenge
      tags:
      - challenges
  /challenges/{id}/invite-link:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get invite link for a challenge
      tags:
      - challenges
  /challenges/{id}/join:
    post:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Join a public challenge
      tags:
      - challenges
...
  /challenges/{id}/progress:
    get:
      parameters:
      - description: Challenge ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Progress'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get user progress in a challenge
      tags:
      - checkin
...
  /challenges/join/{inviteToken}:
    post:
      parameters:
      - description: Invite token
        in: path
        name: inviteToken
        required: true
        type: string
      responses:
        "204":
          description: No Content
      security:
      - BearerAuth: []
      summary: Join challenge by invite token
      tags:
      - challenges
  /challenges/my:
    get:
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Challenge'
            type: array
      security:
      - BearerAuth: []
      summary: List my challenges
      tags:
      - challenges
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"